SPRINT09 - Alterando metodo de consultar PartidaDTO para sempre pegar a proxima partida

+++ src/main/java/com/example/BackEnd/domain/EtapaProducao.java
@@ -23,7 +23,7 @@ public class EtapaProducao {
 	@GeneratedValue(strategy= GenerationType.IDENTITY)
 	private Long id;
 	//d
-	private Long sequencia;
+	private int sequencia;
 	private String inicioPrevisto;
 	private String fimPrevisto;
 	private Long qtdEmEspera;
